Update dependencies and add DOM library to tsconfig
This commit updates axios to redaxios in package.json to benefit from additional features like automatic transforms for JSON data. Also, the version has been updated in package-lock.json and axios replaced with redaxios. The 'follow-redirects' dependency is removed from package-lock.json as it's no longer required. DOM library is added to both the 'tsconfig.json' and 'tsconfig.build.json' files to allow the use of APIs like 'fetch' and 'ReadableStream' in the project.

Moreover, this commit contains updates related to usage of 'ReadableStream' instead of 'Stream' in 'src/typings/yandex.ts' and 'src/main.ts', and removing import of 'URLSearchParams' in 'src/auth.ts' as it's not being used.