Upgrade project and axios package to newer versions
The project is upgraded to version '1.1.2-fetch.3' from '1.1.2-fetch.2'. Also, the axios package used in the project is updated to version '@ponomarevlad/redaxios@^0.5.1-fix.3' from '@ponomarevlad/redaxios@^0.5.1-fix.1'. This is to ensure that we are working with the latest versions and hence have the latest bug fixes and feature updates.